Refactor: break Site Header out into its own component
