+++ src/client/screens/HomeScreen.tsx
@@ -1,14 +1,10 @@
 /** @jsx jsx */
-import { Button as MenuButton, Wrapper, Menu, MenuItem } from "react-aria-menubutton";
-import Avatar from "react-avatar";
 import React, { useEffect, useState } from "react";
 import { connect } from "react-redux";
-import { Link, useHistory } from "react-router-dom";
+import { Link } from "react-router-dom";
 import TimeAgo from "timeago-react";
-import * as H from "history";
 import ProjectListFlyout from "../components/ProjectListFlyout";
 import Icon from "../components/Icon";
-import SupportMenu from "../components/SupportMenu";
 import {
   Alert,
   Box,
@@ -21,115 +17,26 @@ import {
   Styled,
   ThemeUIStyleObject
 } from "theme-ui";
-import { ReactComponent as Logo } from "../media/logos/logo.svg";
 import { ReactComponent as NoMapsIllustration } from "../media/no-maps-illustration.svg";
 
-import { resetState } from "../actions/root";
 import { projectsFetch } from "../actions/projects";
 import { userFetch } from "../actions/user";
 import { resendConfirmationEmail } from "../api";
-import { clearJWT, getJWT } from "../jwt";
+import { getJWT } from "../jwt";
 import { State } from "../reducers";
 import { UserState } from "../reducers/user";
 import { Resource, WriteResource } from "../resource";
 import store from "../store";
 import { IProject } from "../../shared/entities";
 import DeleteProjectModal from "../components/DeleteProjectModal";
+import SiteHeader from "../components/SiteHeader";
 
 interface StateProps {
   readonly projects: Resource<readonly IProject[]>;
   readonly user: UserState;
 }
 
-enum UserMenuKeys {
-  Logout = "logout"
-}
-
-const logout = () => {
-  clearJWT();
-  store.dispatch(resetState());
-};
-
 const style: ThemeUIStyleObject = {
-  header: {
-    alignItems: "center",
-    justifyContent: "space-between",
-    py: 3,
-    px: 3,
-    bg: "gray.0",
-    borderBottom: "1px solid",
-    borderColor: "gray.1",
-    boxShadow: "small"
-  },
-  logoLink: {
-    borderRadius: "small",
-    "&:focus": {
-      outline: "none",
-      boxShadow: "focus"
-    }
-  },
-  avatar: {
-    fontFamily: "heading",
-    cursor: "pointer",
-    ".sb-avatar__text": {
-      "&:hover": {
-        backgroundColor: "#395c78 !important"
-      },
-      "&:active": {
-        backgroundColor: "#2c485e !important"
-      }
-    }
-  },
-  menuButton: {
-    display: "flex",
-    alignItems: "center",
-    bg: "transparent",
-    p: 1,
-    borderRadius: "small",
-    "&:focus": {
-      outline: "none",
-      boxShadow: "focus"
-    }
-  },
-  menu: {
-    width: "150px",
-    position: "absolute",
-    mt: 2,
-    right: 2,
-    bg: "muted",
-    py: 1,
-    px: 1,
-    border: "1px solid",
-    borderColor: "gray.2",
-    boxShadow: "small",
-    borderRadius: "small"
-  },
-  menuList: {
-    p: "0",
-    m: "0",
-    listStyleType: "none"
-  },
-  menuListItem: {
-    borderRadius: "small",
-    py: 1,
-    px: 2,
-    "&:hover:not([disabled])": {
-      bg: "gray.0",
-      cursor: "pointer"
-    },
-    "&[disabled]": {
-      color: "gray.3",
-      cursor: "not-allowed"
-    },
-    "&:focus": {
-      bg: "gray.0",
-      outline: "none",
-      boxShadow: "focus"
-    },
-    "&:active": {
-      bg: "gray.1"
-    }
-  },
   projectRow: {
     textDecoration: "none",
     display: "flex",
@@ -155,7 +62,6 @@ const style: ThemeUIStyleObject = {
 };
 
 const HomeScreen = ({ projects, user }: StateProps) => {
-  const history = useHistory();
   const [resendEmail, setResendEmail] = useState<WriteResource<void, void>>({ data: void 0 });
   const isLoggedIn = getJWT() !== null;
   const projectList =
@@ -212,51 +118,7 @@ const HomeScreen = ({ projects, user }: StateProps) => {
           </Box>
         </Alert>
       )}
-      <Flex as="header" sx={style.header}>
-        <Heading as="h1" sx={{ mb: "0px", mr: "auto", p: 2 }}>
-          <Link to="/" sx={style.logoLink}>
-            <Logo sx={{ width: "15rem" }} />
-          </Link>
-        </Heading>
-        {!isLoggedIn ? (
-          <React.Fragment>
-            <Link to="/login" sx={{ p: 2 }}>
-              Login
-            </Link>{" "}
-            <Link to="/register" sx={{ p: 2 }}>
-              Register
-            </Link>
-          </React.Fragment>
-        ) : "resource" in user ? (
-          <React.Fragment>
-            <SupportMenu />
-            <Wrapper onSelection={handleSelection(history)} sx={{ ml: 3 }}>
-              <MenuButton sx={style.menuButton}>
-                <Avatar
-                  name={user.resource.name}
-                  round={true}
-                  size={"2.5rem"}
-                  color={"#2c485e"}
-                  maxInitials={3}
-                  sx={style.avatar}
-                />
-                <div sx={{ ml: 2 }}>
-                  <Icon name="chevron-down" />
-                </div>
-              </MenuButton>
-              <Menu sx={style.menu}>
-                <ul sx={style.menuList}>
-                  <li key={UserMenuKeys.Logout}>
-                    <MenuItem value={UserMenuKeys.Logout} sx={style.menuListItem}>
-                      Logout
-                    </MenuItem>
-                  </li>
-                </ul>
-              </Menu>
-            </Wrapper>
-          </React.Fragment>
-        ) : null}
-      </Flex>
+      <SiteHeader user={user} />
       <Flex
         as="main"
         sx={{ width: "100%", maxWidth: "large", my: 6, mx: "auto", flexDirection: "column", px: 4 }}
@@ -355,14 +217,6 @@ const HomeScreen = ({ projects, user }: StateProps) => {
       </Flex>
     </Flex>
   );
-};
-
-const handleSelection = (history: H.History) => (key: string | number) => {
-  // eslint-disable-next-line
-  if (key === UserMenuKeys.Logout) {
-    logout();
-    history.push("/login");
-  }
 };
 
 function mapStateToProps(state: State): StateProps {
